# API_REST/swagger.yaml
openapi: 3.0.3
info:
  title: API de Tarefas e Usuários
  description: API REST para gerenciamento de tarefas com autenticação JWT
  version: "1.1.0"

servers:
  - url: http://localhost:3000
    description: Servidor local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Tarefa:
      type: object
      properties:
        id:
          type: string
          example: "64fbcf1e7d3a2c0012345678"
        nome:
          type: string
          example: "Estudar"
        concluida:
          type: boolean
          example: false
      required:
        - id
        - nome
        - concluida

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: "pedroaraujomarcato@gmail.com"
        senha:
          type: string
          example: "abcd1234"
      required:
        - email
        - senha

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    UserRegisterRequest:
      type: object
      properties:
        nome:
          type: string
          example: "Pedro Marcato"
        email:
          type: string
          example: "pedroaraujomarcato@gmail.com"
        senha:
          type: string
          example: "abcd1234"
      required:
        - nome
        - email
        - senha

    UserResponse:
      type: object
      properties:
        id:
          type: string
          example: "671a0e88..."
        nome:
          type: string
          example: "Pedro Marcato"
        email:
          type: string
          example: "pedroaraujomarcato@gmail.com"

paths:
  /users/register:
    post:
      tags:
        - Users
      summary: Cadastro de novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: E-mail já cadastrado ou dados inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "E-mail já cadastrado"

  /users/login:
    post:
      tags:
        - Users
      summary: Autenticação do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token JWT gerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Credenciais inválidas"

  /users/renovar:
    post:
      tags:
        - Users
      summary: Renovar token JWT
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Novo token JWT gerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Token inválido"

  /tarefas:
    get:
      tags:
        - Tarefas
      summary: Lista todas as tarefas
      responses:
        '200':
          description: Lista de tarefas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tarefa'

    post:
      tags:
        - Tarefas
      summary: Cria uma nova tarefa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  example: "Estudar"
              required:
                - nome
      responses:
        '201':
          description: Tarefa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
        '422':
          description: Erro de validação
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Nome da tarefa é obrigatório"

  /tarefas/{id}:
    get:
      tags:
        - Tarefas
      summary: Retorna uma tarefa pelo ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID da tarefa
      responses:
        '200':
          description: Tarefa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "ID inválido"
        '404':
          description: Tarefa não encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Tarefa não encontrada"

    put:
      tags:
        - Tarefas
      summary: Atualiza uma tarefa pelo ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID da tarefa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  example: "Estudar REST"
                concluida:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Tarefa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
        '400':
          description: ID inválido
        '404':
          description: Tarefa não encontrada
        '422':
          description: Erro de validação

    delete:
      tags:
        - Tarefas
      summary: Deleta uma tarefa pelo ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID da tarefa
      responses:
        '204':
          description: Tarefa deletada
        '400':
          description: ID inválido
        '404':
          description: Tarefa não encontrada
          